Get the source closer to cc9cii branch for eventual release of NIF code.  Lots of old NIF code as well as comments were removed.  Vertex animation works, if jerky.

+++ apps/openmw/mwdialogue/dialoguemanagerimp.cpp
@@ -248,7 +248,7 @@ namespace MWDialogue
                 MWScript::InterpreterContext interpreterContext(&mActor.getRefData().getLocals(),mActor);
                 Interpreter::Interpreter interpreter;
                 MWScript::installOpcodes (interpreter);
-                interpreter.run (&code[0], static_cast<int>(code.size()), interpreterContext);
+                interpreter.run (&code[0], code.size(), interpreterContext);
             }
             catch (const std::exception& error)
             {
